Community gas optimization, and alternate darwin contract created
Made the community contract more gas efficient when checking if votes are valid, and in deleting the votes array. Also made an alternate version of the darwin token contract that reads currentRate as a state variable, that gets updated whenever the reflection address is send funds. This ideally removes a lot of the gas costs of checking the rate on each transfer as it only should get updated on reflection. Also in the darwin alt contract some of the values that weren't set after the constructor were changed to constants, with some notices about some variables that are never set at all.