Introduced smartSync
SmartSync is a settable boolean (by the DEFAULT_ADMIN_ROLE). Default: false.

When it is false, all pairs of darwin are synced on all transfers that are not buys. (spends more gas but grants more stability in the LP syncing since it is also called on normal transfers between 2 EOAs)

When it is true, only the pair that is interested by the current transfer is synced, and only when the transfer is a sell. (spends less gas but grants less stability to LP syncing)