Add option to identify a router/pair as darwinswap
When setting a new router, now the owner can also use a bool to tell "if it's darwinswap or not". If the bool is true, the router and pair addresses will be added to a mapping that, when doing a transfer, will be checked in _beforeTokenTransfer. If we're trading on darwinswap, the Tokenomics 2.0 logic of the darwin token won't be called (cause DarwinSwap itself will work them out).