Added startPresale() function
Since we don't know exactly when we will start presale but we will deploy its contract together with Darwin Protocol's, I removed the presaleStart setting from the init function and created a function to start it as soon as it's executed. Also removed the router from the arguments because we're going to deploy darwinswap AFTER we deploy the presale contract.

Also, in the Darwin Protocol contract I added a function "setPrivateSaleAddress" if we're going to do more than 1 private sales.